Unignore Cargo.lock and remove duplicate entries in .gitignore
According to both the comment and its linked docs page, the lock file
should be left out of the .gitignore in executables, so I removed it and
checked it back into git after building.

I also removed the duplicate entries.

(this will also make building the crate with Nix easier, as it requires
a Cargo.lock to help ensure the build is reproducible)